+++ example/bounce/view/board.dart
@@ -1,0 +1,24 @@
+part of bounce;
+
+class Board extends Surface {
+  var movablePieces = new MovablePieces(32);
+  
+  Board(CanvasElement canvas) : super(canvas) {
+    movablePieces.randomInit();
+    //movablePieces.forEach((MovablePiece mp) => mp.shape = PieceShape.CIRCLE);
+    window.animationFrame.then(gameLoop); 
+  }
+  
+  gameLoop(num delta) {
+    draw();
+    window.animationFrame.then(gameLoop);
+  }
+  
+  draw() {
+    clear();
+    movablePieces.forEach((MovablePiece mp) {
+      mp.move();
+      new Circle(this, mp.x, mp.y, mp.width / 2, color: mp.colorCode).draw();
+    });
+  }
+}